Rename run-karma to karma-service and add more tests
