+++ lib/commands/test/karma.js
@@ -1,7 +1,7 @@
 const importLazy = require('import-lazy')(require);
 
 const { mainHandler } = importLazy('../../cli/main-handler');
-const { runKarmaTests } = importLazy('../../test/run-karma');
+const KarmaService = importLazy('../../test/karma-service');
 const StripesPlatform = importLazy('../../platform/stripes-platform');
 const { applyOptions, serverOptions, okapiOptions, stripesConfigOptions } = importLazy('../common-options');
 const StripesCore = importLazy('../../cli/stripes-core');
@@ -27,7 +27,9 @@ function karmaCommand(argv, context) {
   console.log('Starting Karma tests...');
   const stripes = new StripesCore(context, platform.aliases);
   const webpackConfig = stripes.getStripesWebpackConfig(platform.getStripesConfig(), Object.assign({}, argv, { webpackOverrides }));
-  runKarmaTests(webpackConfig, argv.karma);
+
+  const karmaService = new KarmaService(context.cwd);
+  karmaService.runKarmaTests(webpackConfig, argv.karma);
 }
 
 module.exports = {
